+++ lib/src/application/application.dart
@@ -3,6 +3,7 @@
 /// APIGuide package
 
 /// Code starts here
+export '../../src/application/helpers/functions.dart';
 export '../../src/application/sample_code.dart';
 export '../../src/application/theme_state.dart';